Altera atributos de sharedViewModels para ULong e os usos desses atributos em outras classes para suportar o tipo

+++ app/src/main/java/com/example/calculadorafatoriais/SharedViewModel.kt
@@ -3,6 +3,6 @@ package com.example.calculadorafatoriais
 import androidx.lifecycle.ViewModel
 
 class SharedViewModel : ViewModel() {
-    var n: Int? = null
-    var k: Int? = null
+    var n: ULong? = null
+    var k: ULong? = null
 }
